Erste Funktionen über DB ermöglicht
User hinzufügen und löschen in SQLite3 Datenbank

+++ prototyp2/user.php
@@ -1,7 +1,10 @@
-<!-- um nach 3 Sekunden zur form.php zurück zu springen -->
+<!-- um nach 3 Sekunden zur form.php zurück zu springen 
 <meta  http-equiv="refresh" content="2; URL=form.php" http-equiv="content-type" content="text/html; charset=utf-8" />
-		
+		-->
+
 <?php
+require_once('db.php');
+
 $file = "../passwd/.htpasswd";
 $file_new = "../passwd/.htpasswd.new"; //Datei zum zwischenspeichern für Löschfunktion
 $mode = "a+";
@@ -10,6 +13,9 @@ $name_ok = true;
 $pwd1_ok = true;
 $pwd2_ok =true;
 $radio_check = $_POST['radio_tbl'];
+
+$group = 'user';
+$mail = 'test@test.de';
 
 switch($radio_check){
 	
@@ -27,72 +33,44 @@ switch($radio_check){
 		echo "Bitte Passwort wiederholen";
 		$pwd2_ok = false;
 	}
+	if ($_POST['new_pwd1'] != $_POST['new_pwd2'])
+			die("Die Passwörter stimmen nicht überein");
 	
 	//Neuen Benutzer eintragen
 	if($name_ok && $pwd1_ok && $pwd2_ok){
-		$htpwfile = fopen($file, $mode);
-		if ($_POST['new_pwd1'] != $_POST['new_pwd2']){
-			die("Die Passwörter stimmen nicht überein");
-		}
-		if($htpwfile==NULL){
-			echo "Datei konnte nicht gelesen werden";
-		}else {
-			//überprüfen ob Benutzer schon vorhanden
-			$length = sizeof($lines);
-			for($i=0;$i<$length;$i++){
-				$user_list = explode(":", $lines[$i]);
-			}
-			$pattern = $_POST['new_name'];
-			if(in_array($pattern, $user_list)){
-				die( "Benutzer ist schon vorhanden");//bricht Vorgang ab und erstellt Ausgabe
-			}
-			//Benutzer ist nicht vorhanden, also wird er hinzugefügt
-			if(fwrite($htpwfile, $_POST['new_name'].":".$_POST['new_pwd1']."\r\n")){
-				echo "Neuer Benutzer ".$_POST['new_name']." wurde hinzugef&uuml;gt. <br>";
-			}
-			if(!fclose($htpwfile)){
-				echo "Fehler";
-			}
-		}
+		$db = new SQLite3('db/user.db');
+		
+		if(!$db)
+			die($db->lastErrorMsg());
+			$name = $_POST['new_name'];
+			$pwd = $_POST['new_pwd1'];
+		$result=$db->querySingle("Select Benutzername from user where Benutzername='$name'");
+		
+		if($result){
+			echo "Benutzername ist schon vorhanden";
+		}else{		
+				$db->exec("INSERT INTO user(Benutzername, Passwort, Gruppe, EMail)
+				VALUES ('$name', '$pwd', '$group', '$mail');");
+				echo "Benutzer wurde hinzugef&uuml;gt";		
+}
+		$db->close();
 	}
 	break;
 	
 	case 'clr':
-	if($_POST['user']==""){
-		die("Keinen Benutzernamen eingegeben! <br>");
-	}
-	$pattern = $_POST['user']; //pattern für die Suche nach dem Benutzernamen
-	$length = sizeof($lines);
-	$a = 0;
-	//erstellen eines neuen Array, mit den Usern die weiterhin zugriff haben sollen
-	for($i=0; $i<$length; $i++){
-		$user = explode(":", $lines[$i]);
-		if($user[0]!=$pattern){
-			$user_list[$a] = $lines[$i];
-			$a++;
+		if($_POST['user']==""){
+			die("Keinen Benutzernamen eingegeben! <br>");
 		}
-	}
-	if($length==$a)
-		die("Benutzer wurde nicht gefunden!");
-	
-	//zurückschreiben der user:passwort kombination
-	$htpwfile = fopen($file_new, $mode);
-	if($htpwfile==NULL){
-		echo "Datei konnte nicht gelesen werden";
-	}else {
-		$length = sizeof($user_list);
-		for($i=0;$i<$length;$i++)
-			fwrite($htpwfile, $user_list[$i]);
-		if(!fclose($htpwfile))
-				echo "Fehler";
-		if(file_exists($file)==true){
-			if(unlink($file)==true){
-				if(rename("../passwd/.htpasswd.new", "../passwd/.htpasswd")==true)
-					echo "Benutzer ".$_POST['user']." wurde gel&ouml;scht. <br>";
-			}
+		$db = new SQLite3('db/user.db');
+		if(!$db)
+			die($db->lastErrorMsg());
+		$name = $_POST['user'];
+		$result = $db->querySingle("DELETE FROM user where Benutzername='$name'");
+		if($result){
+			echo "Eintrag konnte nicht geschrieben werden";
 		}
-	}
-	break;
+		$db->close();
+		break;
 	
 	case 'pwd_change':
 	if($_POST['chg_name']==""){
@@ -108,40 +86,42 @@ switch($radio_check){
 		echo "Bitte Passwort wiederholen";
 		$pwd2_ok = false;
 	}
-	if ($_POST['chg_pwd1'] != $_POST['chg_pwd2']){
-		die("Die Passwörter stimmen nicht überein");
-	}
-	$pattern = $_POST['chg_name']; //pattern für die Suche nach dem Benutzernamen
-	$length = sizeof($lines);
-	$a = 0;
-	//erstellen eines neuen Array, mit den Usern die weiterhin zugriff haben sollen
-	for($i=0; $i<$length; $i++){
-		$user = explode(":", $lines[$i]);
-		if($user[0]!=$pattern){
-			$user_list[$a] = $lines[$i];
-		}else{
-			$user_list[$a] = $user[0].":".$_POST['chg_pwd1']."\r\n";
-			break;
+	if($name_ok && $pwd1_ok && $pwd2_ok){
+		if ($_POST['chg_pwd1'] != $_POST['chg_pwd2']){
+			die("Die Passwörter stimmen nicht überein");
 		}
-		$a++;
-	}
-	if($length==$a)
-		die("Benutzer wurde nicht gefunden!");
-	
-	//zurückschreiben der user:passwort kombination
-	$htpwfile = fopen($file_new, $mode);
-	if($htpwfile==NULL){
-		echo "Datei konnte nicht gelesen werden";
-	}else {
-		$length = sizeof($user_list);
-		for($i=0;$i<$length;$i++)
-			fwrite($htpwfile, $user_list[$i]);
-		if(!fclose($htpwfile))
-				echo "Fehler";
-		if(file_exists($file)==true){
-			if(unlink($file)==true){
-				if(rename("../passwd/.htpasswd.new", "../passwd/.htpasswd")==true)
-					echo "Passwort f&uuml;r Benutzer ".$_POST['chg_name']." wurde gel&auml;ndert. <br>";
+		$pattern = $_POST['chg_name']; //pattern für die Suche nach dem Benutzernamen
+		$length = sizeof($lines);
+		$a = 0;
+		//erstellen eines neuen Array, mit den Usern die weiterhin zugriff haben sollen
+		for($i=0; $i<$length; $i++){
+			$user = explode(":", $lines[$i]);
+			if($user[0]!=$pattern){
+				$user_list[$a] = $lines[$i];
+			}else{
+				$user_list[$a] = $user[0].":".$_POST['chg_pwd1']."\r\n";
+				break;
+			}
+			$a++;
+		}
+		if($length==$a)
+			die("Benutzer wurde nicht gefunden!");
+		
+		//zurückschreiben der user:passwort kombination
+		$htpwfile = fopen($file_new, $mode);
+		if($htpwfile==NULL){
+			echo "Datei konnte nicht gelesen werden";
+		}else {
+			$length = sizeof($user_list);
+			for($i=0;$i<$length;$i++)
+				fwrite($htpwfile, $user_list[$i]);
+			if(!fclose($htpwfile))
+					echo "Fehler";
+			if(file_exists($file)==true){
+				if(unlink($file)==true){
+					if(rename("../passwd/.htpasswd.new", "../passwd/.htpasswd")==true)
+						echo "Passwort f&uuml;r Benutzer ".$_POST['chg_name']." wurde gel&auml;ndert. <br>";
+				}
 			}
 		}
 	}
